pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "openai>=1.61.0",
    "python-dotenv>=1.0.1",
]
